feat: Add quick status dropdown with color indicators and smooth animations
- Added quick status dropdown in left header (Available, Break, Lunch, EOS only)
- Implemented color-coded status indicators (green, orange, blue, red)
- Added smooth animations and hover effects
- Improved UI spacing and structure for cleaner design
- Enhanced dropdown with custom styling and transitions

+++ Models/Agent.cs
@@ -28,7 +28,7 @@ namespace NextHorizon.Models
         public string PreviewQuestion { get; set; } = string.Empty;
 
         [Required]
-        public byte ChatSlot { get; set; }
+        public int ChatSlot { get; set; }
 
         [Required]
         [MaxLength(20)]
